Add `journal` migration logic and tests. Implement `MigrateJournal` for legacy schedule and reply data. Introduce repair request classification, placeholder author handling, and detailed migration reporting. Update `db-migration-strategist` documentation with `journal` context.

+++ internal/application/property/create_property.go
@@ -16,7 +16,7 @@ import (
 type CreatePropertyInput struct {
 	Name                 string
 	Address              string
-	ElectricityUnitPrice int
+	ElectricityUnitPrice float64
 	OwnerID              string
 }
